requirements.txt
# ==================================
# UNSW-NB15 Imbalance Study
# Python Dependencies
# ==================================

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0

# Development (optional)
jupyter>=1.0.0
